Move flowrgui screenshot tests to separate file for codecov exclusion
- Extract screenshot tests from mod test in main.rs to a separate
  screenshot_test.rs module (same pattern as flowedit)
- Make test_gui() pub(super) so screenshot_test module can access it
- Add flowr/src/bin/flowrgui/screenshot_test.rs to codecov ignore list
- Remove PNGs from PR branch (CI regenerates them)
